+++ src/http/sendfile.c
@@ -154,9 +154,11 @@ int httpd_pipe_set_chunked (httpd_client_t * http, hin_pipe_t * pipe) {
 }
 
 int httpd_pipe_upload_chunked (httpd_client_t * http, hin_pipe_t * pipe) {
+  if (http->disable & HIN_HTTP_CHUNKUP) return 0;
   if (http->peer_flags & HIN_HTTP_CHUNKUP) {
     pipe->decode_callback = hin_pipe_decode_chunked;
     pipe->read_callback = hin_pipe_copy_chunked;
   }
+  return 0;
 }