+++ src/lua/lua_config.c
@@ -11,22 +11,29 @@
 #include "conf.h"
 
 static int l_hin_create_httpd (lua_State *L) {
-  int ref1 = 0, ref2 = 0;
+  int ref1 = 0, ref2 = 0, ref3 = 0;
   if (lua_type (L, 1) != LUA_TFUNCTION) {
     printf ("not provided a callback\n");
     return 0;
   }
+
   lua_pushvalue (L, 1);
   ref1 = luaL_ref (L, LUA_REGISTRYINDEX);
 
-  if (lua_type (L, 1) == LUA_TFUNCTION) {
+  if (lua_type (L, 2) == LUA_TFUNCTION) {
     lua_pushvalue (L, 2);
     ref2 = luaL_ref (L, LUA_REGISTRYINDEX);
   }
 
+  if (lua_type (L, 3) == LUA_TFUNCTION) {
+    lua_pushvalue (L, 3);
+    ref3 = luaL_ref (L, LUA_REGISTRYINDEX);
+  }
+
   hin_server_data_t * server = calloc (1, sizeof (hin_server_data_t));
   server->request_callback = ref1;
-  server->finish_callback = ref2;
+  server->error_callback = ref2;
+  server->finish_callback = ref3;
   server->L = L;
   server->magic = HIN_SERVER_MAGIC;
   server->timeout = HIN_HTTPD_TIMEOUT;